+++ crates/acts/src/cache/tests.rs
@@ -1,5 +1,5 @@
 use crate::{
-    EngineBuilder, Workflow, data,
+    Engine, Workflow, data,
     scheduler::{NodeTree, Process, TaskState},
     store::DbCollectionIden,
     utils,
@@ -7,7 +7,7 @@ use crate::{
 
 #[tokio::test]
 async fn cache_count() {
-    let engine = EngineBuilder::new().cache_size(10).build().start().unwrap();
+    let engine = Engine::builder().cache_size(10).build().start().unwrap();
     let rt = engine.runtime();
     let cache = rt.cache();
 
@@ -18,7 +18,7 @@ async fn cache_count() {
 
 #[tokio::test]
 async fn cache_push_get() {
-    let engine = EngineBuilder::new().cache_size(10).build().start().unwrap();
+    let engine = Engine::builder().cache_size(10).build().start().unwrap();
     let rt = engine.runtime();
     let cache = rt.cache();
     let pid = utils::longid();
@@ -32,7 +32,7 @@ async fn cache_push_get() {
 
 #[tokio::test]
 async fn cache_push_to_store() {
-    let engine = EngineBuilder::new().cache_size(1).build().start().unwrap();
+    let engine = Engine::builder().cache_size(1).build().start().unwrap();
     let rt = engine.runtime();
     let cache = rt.cache();
 
@@ -53,7 +53,7 @@ async fn cache_push_to_store() {
 
 #[tokio::test]
 async fn cache_remove() {
-    let engine = EngineBuilder::new().cache_size(10).build().start().unwrap();
+    let engine = Engine::builder().cache_size(10).build().start().unwrap();
     let rt = engine.runtime();
     let cache = rt.cache();
 
@@ -81,7 +81,7 @@ async fn cache_remove() {
 
 #[tokio::test]
 async fn cache_upsert() {
-    let engine = EngineBuilder::new().cache_size(10).build().start().unwrap();
+    let engine = Engine::builder().cache_size(10).build().start().unwrap();
     let rt = engine.runtime();
     let mut workflow = Workflow::new().with_step(|step| step.with_name("step1"));
 
@@ -105,7 +105,7 @@ async fn cache_upsert() {
 
 #[tokio::test]
 async fn cache_restore_count() {
-    let engine = EngineBuilder::new().cache_size(5).build().start().unwrap();
+    let engine = Engine::builder().cache_size(5).build().start().unwrap();
     let model = Workflow::new()
         .with_id("m1")
         .with_step(|step| step.with_name("step1"));
@@ -137,7 +137,7 @@ async fn cache_restore_count() {
 
 #[tokio::test]
 async fn cache_restore_working_state() {
-    let engine = EngineBuilder::new().cache_size(5).build().start().unwrap();
+    let engine = Engine::builder().cache_size(5).build().start().unwrap();
     let model = Workflow::new()
         .with_id("m1")
         .with_step(|step| step.with_name("step1"));
@@ -182,7 +182,7 @@ async fn cache_restore_working_state() {
 
 #[tokio::test]
 async fn cache_restore_completed_state() {
-    let engine = EngineBuilder::new().cache_size(5).build().start().unwrap();
+    let engine = Engine::builder().cache_size(5).build().start().unwrap();
     let model = Workflow::new()
         .with_id("m1")
         .with_step(|step| step.with_name("step1"));
@@ -227,7 +227,7 @@ async fn cache_restore_completed_state() {
 
 #[tokio::test]
 async fn cache_restore_less_cap() {
-    let engine = EngineBuilder::new().cache_size(5).build().start().unwrap();
+    let engine = Engine::builder().cache_size(5).build().start().unwrap();
     let model = Workflow::new()
         .with_id("m1")
         .with_step(|step| step.with_name("step1"));