+++ examples/user_var/main.rs
@@ -1,11 +1,11 @@
 mod module;
 mod plugin;
 
-use acts::{EngineBuilder, Result, Vars, Workflow};
+use acts::{Engine, Result, Vars, Workflow};
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    let engine = EngineBuilder::new()
+    let engine = Engine::builder()
         .add_plugin(&plugin::UserVarPlugin)
         .build()
         .start()?;