+++ CHANGELOG.md
@@ -188,3 +188,4 @@
 - use `options` to support the user custom extension, such as `tag`, `rn`
 - add `exposes` to workflow, step, branch and act instead of options.exposes
 - change `timeouts` and `catches` to support complex flow like `steps`
+- add `Engine::builder()` to `Engine` and remove the export of EngineBuilder